+++ universal.js
@@ -4,6 +4,7 @@ module.exports = {
         "eslint-config-konnektid/envs/universal",
         "eslint-config-konnektid/envs/webpack",
         "eslint-config-konnektid/rules/base",
+        "eslint-config-konnektid/rules/babel",
         "eslint-config-konnektid/rules/universal"
     ]
 };